Add a workaround quest marker for Dark Brotherhood gem
Add a quest marker that points to the Dark Brotherhood door as long as
it's disabled. The door is enabled when you first make it into the
sanctuary during DB02.

Before this point, the in-game door is replaced by an activator. The
activator is disabled when you make it into the sanctuary. The marker is
placed on the door instead of the activator for two reasons:

1. It's marginally more correct-looking. On the world map, the marker
   shows the door quest marker icon. In world and on the compass, it
   still shows the regular marker icon. The activator shows the regular
   icon even on the world map.
2. For some reason, when DB02 hits stage 40, a quest marker on the door
   activator bugs on the world map, appearing as three markers
   overlapping each other. The door doesn't have this problem.